+++ magidoc.mjs
@@ -37,6 +37,7 @@ To learn how to deactivate it, read the
         'Polyline': '<>',
         'GeoJson': '<>',
         'OffsetDateTime': '2024-02-05T18:04:23+01:00',
+        'LocalDate': '2024-05-24',
         'Duration': 'PT10M',
         'CoordinateValue': 19.24,
         'Reluctance': 3.1,
@@ -44,7 +45,6 @@ To learn how to deactivate it, read the
         'Cost': 100,
         'Ratio': 0.25,
         'Locale': 'en'
-
       },
     }
   },