+++ src/ext/java/org/opentripplanner/ext/siri/SiriTimetableSnapshotSource.java
@@ -57,7 +57,12 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
    */
   private final SiriTripPatternCache tripPatternCache;
 
-  private final TransitEditorService transitService;
+  /**
+   * Long-lived transit editor service that has access to the timetable snapshot buffer.
+   * This differs from the usual use case where the transit service refers to the latest published
+   * timetable snapshot.
+   */
+  private final TransitEditorService transitEditorService;
 
   private final TimetableSnapshotManager snapshotManager;
 
@@ -71,9 +76,10 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
         parameters,
         () -> LocalDate.now(transitModel.getTimeZone())
       );
-    this.transitService = new DefaultTransitService(transitModel);
+    this.transitEditorService =
+      new DefaultTransitService(transitModel, getTimetableSnapshotBuffer());
     this.tripPatternCache =
-      new SiriTripPatternCache(tripPatternIdGenerator, transitService::getPatternForTrip);
+      new SiriTripPatternCache(tripPatternIdGenerator, transitEditorService::getPatternForTrip);
 
     transitModel.initTimetableSnapshotProvider(this);
   }
@@ -102,26 +108,22 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
 
     List<Result<UpdateSuccess, UpdateError>> results = new ArrayList<>();
 
-    snapshotManager.withLock(() -> {
-      if (incrementality == FULL_DATASET) {
-        // Remove all updates from the buffer
-        snapshotManager.clearBuffer(feedId);
-      }
+    if (incrementality == FULL_DATASET) {
+      // Remove all updates from the buffer
+      snapshotManager.clearBuffer(feedId);
+    }
 
-      for (var etDelivery : updates) {
-        for (var estimatedJourneyVersion : etDelivery.getEstimatedJourneyVersionFrames()) {
-          var journeys = estimatedJourneyVersion.getEstimatedVehicleJourneies();
-          LOG.debug("Handling {} EstimatedVehicleJourneys.", journeys.size());
-          for (EstimatedVehicleJourney journey : journeys) {
-            results.add(apply(journey, transitService, fuzzyTripMatcher, entityResolver));
-          }
+    for (var etDelivery : updates) {
+      for (var estimatedJourneyVersion : etDelivery.getEstimatedJourneyVersionFrames()) {
+        var journeys = estimatedJourneyVersion.getEstimatedVehicleJourneies();
+        LOG.debug("Handling {} EstimatedVehicleJourneys.", journeys.size());
+        for (EstimatedVehicleJourney journey : journeys) {
+          results.add(apply(journey, transitEditorService, fuzzyTripMatcher, entityResolver));
         }
       }
+    }
 
-      LOG.debug("message contains {} trip updates", updates.size());
-
-      snapshotManager.purgeAndCommit();
-    });
+    LOG.debug("message contains {} trip updates", updates.size());
 
     return UpdateResult.ofResults(results);
   }
@@ -129,6 +131,10 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
   @Override
   public TimetableSnapshot getTimetableSnapshot() {
     return snapshotManager.getTimetableSnapshot();
+  }
+
+  private TimetableSnapshot getTimetableSnapshotBuffer() {
+    return snapshotManager.getTimetableSnapshotBuffer();
   }
 
   private Result<UpdateSuccess, UpdateError> apply(
@@ -195,11 +201,7 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
    * Snapshot timetable is used as source if initialised, trip patterns scheduled timetable if not.
    */
   private Timetable getCurrentTimetable(TripPattern tripPattern, LocalDate serviceDate) {
-    TimetableSnapshot timetableSnapshot = getTimetableSnapshot();
-    if (timetableSnapshot != null) {
-      return timetableSnapshot.resolve(tripPattern, serviceDate);
-    }
-    return tripPattern.getScheduledTimetable();
+    return getTimetableSnapshotBuffer().resolve(tripPattern, serviceDate);
   }
 
   private Result<TripUpdate, UpdateError> handleModifiedTrip(
@@ -228,7 +230,7 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
 
     if (trip != null) {
       // Found exact match
-      pattern = transitService.getPatternForTrip(trip);
+      pattern = transitEditorService.getPatternForTrip(trip);
     } else if (fuzzyTripMatcher != null) {
       // No exact match found - search for trips based on arrival-times/stop-patterns
       TripAndPattern tripAndPattern = fuzzyTripMatcher.match(
@@ -263,7 +265,7 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
       pattern,
       estimatedVehicleJourney,
       serviceDate,
-      transitService.getTimeZone(),
+      transitEditorService.getTimeZone(),
       entityResolver
     )
       .build();
@@ -310,7 +312,7 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
   private boolean markScheduledTripAsDeleted(Trip trip, final LocalDate serviceDate) {
     boolean success = false;
 
-    final TripPattern pattern = transitService.getPatternForTrip(trip);
+    final TripPattern pattern = transitEditorService.getPatternForTrip(trip);
 
     if (pattern != null) {
       // Mark scheduled trip times for this trip in this pattern as deleted
@@ -328,5 +330,12 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
     }
 
     return success;
+  }
+
+  /**
+   * Flush pending changes in the timetable snapshot buffer and publish a new snapshot.
+   */
+  public void flushBuffer() {
+    snapshotManager.purgeAndCommit();
   }
 }