+++ src/main/java/org/opentripplanner/framework/graphql/scalar/DateScalarFactory.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.apis.transmodel.model.scalars;
+package org.opentripplanner.framework.graphql.scalar;
 
 import graphql.language.StringValue;
 import graphql.schema.Coercing;
@@ -9,21 +9,40 @@ import graphql.schema.GraphQLScalarType;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
+import javax.annotation.Nullable;
 
 public class DateScalarFactory {
 
-  private static final String DOCUMENTATION =
+  private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ISO_LOCAL_DATE;
+  public static final String TRANSMODEL_DESCRIPTION =
     "Local date using the ISO 8601 format: `YYYY-MM-DD`. Example: `2020-05-17`.";
 
-  private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ISO_LOCAL_DATE;
+  private static final String TRANSMODEL_NAME = "Date";
+  private static final String GTFS_NAME = "LocalDate";
 
   private DateScalarFactory() {}
 
-  public static GraphQLScalarType createDateScalar() {
+  public static GraphQLScalarType createTransmodelDateScalar() {
+    return createDateScalar(TRANSMODEL_NAME, TRANSMODEL_DESCRIPTION);
+  }
+
+  public static GraphQLScalarType createGtfsDateScalar() {
+    // description comes from schema.graphqls
+    return createDateScalar(GTFS_NAME, null);
+  }
+
+  /**
+   * @param description Nullable description that allows caller to pass in null which leads to the
+   *                    description from schema.graphqls to be used.
+   */
+  private static GraphQLScalarType createDateScalar(
+    String scalarName,
+    @Nullable String description
+  ) {
     return GraphQLScalarType
       .newScalar()
-      .name("Date")
-      .description(DOCUMENTATION)
+      .name(scalarName)
+      .description(description)
       .coercing(
         new Coercing<LocalDate, String>() {
           @Override
@@ -33,7 +52,7 @@ public class DateScalarFactory {
             }
 
             throw new CoercingSerializeException(
-              "Only LocalDate is supported to serialize but found " + input
+              "Only %s is supported to serialize but found %s".formatted(scalarName, input)
             );
           }
 
@@ -43,7 +62,7 @@ public class DateScalarFactory {
               return LocalDate.from(FORMATTER.parse((String) input));
             } catch (DateTimeParseException e) {
               throw new CoercingParseValueException(
-                "Expected type 'Date' but was '" + input + "'."
+                "Expected type '%s' but was '%s'.".formatted(scalarName, input)
               );
             }
           }