+++ src/main/java/org/opentripplanner/standalone/configure/ConstructApplication.java
@@ -5,7 +5,6 @@ import javax.annotation.Nullable;
 import org.opentripplanner.apis.transmodel.TransmodelAPI;
 import org.opentripplanner.datastore.api.DataSource;
 import org.opentripplanner.ext.emissions.EmissionsDataModel;
-import org.opentripplanner.ext.geocoder.LuceneIndex;
 import org.opentripplanner.ext.stopconsolidation.StopConsolidationRepository;
 import org.opentripplanner.framework.application.LogMDCSupport;
 import org.opentripplanner.framework.application.OTPFeature;
@@ -33,6 +32,7 @@ import org.opentripplanner.standalone.server.GrizzlyServer;
 import org.opentripplanner.standalone.server.OTPWebApplication;
 import org.opentripplanner.street.model.StreetLimitationParameters;
 import org.opentripplanner.street.model.elevation.ElevationUtils;
+import org.opentripplanner.transit.service.DefaultTransitService;
 import org.opentripplanner.transit.service.TransitModel;
 import org.opentripplanner.updater.configure.UpdaterConfigurator;
 import org.opentripplanner.visualizer.GraphVisualizer;
@@ -180,8 +180,9 @@ public class ConstructApplication {
     }
 
     if (OTPFeature.SandboxAPIGeocoder.isOn()) {
-      LOG.info("Creating debug client geocoder lucene index");
-      LuceneIndex.forServer(createServerContext());
+      LOG.info("Initializing geocoder");
+      // eagerly initialize the geocoder
+      this.factory.luceneIndex();
     }
   }
 
@@ -202,7 +203,7 @@ public class ConstructApplication {
     TransitModel transitModel,
     TransitTuningParameters tuningParameters
   ) {
-    if (!transitModel.hasTransit() || transitModel.getTransitModelIndex() == null) {
+    if (!transitModel.hasTransit() || !transitModel.isIndexed()) {
       LOG.warn(
         "Cannot create Raptor data, that requires the graph to have transit data and be indexed."
       );
@@ -211,10 +212,7 @@ public class ConstructApplication {
     transitModel.setTransitLayer(TransitLayerMapper.map(tuningParameters, transitModel));
     transitModel.setRealtimeTransitLayer(new TransitLayer(transitModel.getTransitLayer()));
     transitModel.setTransitLayerUpdater(
-      new TransitLayerUpdater(
-        transitModel,
-        transitModel.getTransitModelIndex().getServiceCodesRunningForDate()
-      )
+      new TransitLayerUpdater(new DefaultTransitService(transitModel))
     );
   }