+++ src/main/java/org/opentripplanner/transit/service/DefaultTransitService.java
@@ -9,6 +9,7 @@ import java.time.LocalDate;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -76,6 +77,14 @@ public class DefaultTransitService implements TransitEditorService {
   public DefaultTransitService(TransitModel transitModel) {
     this.transitModel = transitModel;
     this.transitModelIndex = transitModel.getTransitModelIndex();
+  }
+
+  public DefaultTransitService(
+    TransitModel transitModel,
+    TimetableSnapshot timetableSnapshotBuffer
+  ) {
+    this(transitModel);
+    this.timetableSnapshot = timetableSnapshotBuffer;
   }
 
   @Override
@@ -595,6 +604,16 @@ public class DefaultTransitService implements TransitEditorService {
   }
 
   @Override
+  public void setRealtimeTransitLayer(TransitLayer realtimeTransitLayer) {
+    transitModel.setRealtimeTransitLayer(realtimeTransitLayer);
+  }
+
+  @Override
+  public boolean hasRealtimeTransitLayer() {
+    return transitModel.hasRealtimeTransitLayer();
+  }
+
+  @Override
   public CalendarService getCalendarService() {
     return this.transitModel.getCalendarService();
   }
@@ -657,6 +676,16 @@ public class DefaultTransitService implements TransitEditorService {
   @Override
   public Deduplicator getDeduplicator() {
     return transitModel.getDeduplicator();
+  }
+
+  @Override
+  public Set<LocalDate> getAllServiceCodes() {
+    return Collections.unmodifiableSet(transitModelIndex.getServiceCodesRunningForDate().keySet());
+  }
+
+  @Override
+  public Map<LocalDate, TIntSet> getServiceCodesRunningForDate() {
+    return Collections.unmodifiableMap(transitModelIndex.getServiceCodesRunningForDate());
   }
 
   /**